+++ KinderGarten/Controllers/GroupsController.cs
@@ -13,15 +13,22 @@ namespace KinderGarten.Controllers
     public class GroupsController : Controller
     {
         // GET: Activities
-        public ActionResult Index()
-        {
-            return View();
-        }
+        //public ActionResult Index()
+        //{
+        //    return View();
+        //}
 
         [HttpGet]
         public ActionResult Create()
         {
-            return View();
+            if (User.Identity.IsAuthenticated)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         [HttpPost]
@@ -30,16 +37,24 @@ namespace KinderGarten.Controllers
             UnitOfWork.UOW.GroupRepository.Create(group);
             UnitOfWork.UOW.Save();
             return RedirectToAction(nameof(Read));
-
         }
+
+        
 
 
 
         [HttpGet]
         public ActionResult Read()
         {
-            List<Group> allGroups = UnitOfWork.UOW.GroupRepository.Read();
+            if (User.Identity.IsAuthenticated)
+            {
+                List<Group> allGroups = UnitOfWork.UOW.GroupRepository.Read();
             return View(allGroups);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
 
@@ -47,8 +62,15 @@ namespace KinderGarten.Controllers
         [HttpGet]
         public ActionResult Update(int id)
         {
-            Group group = UnitOfWork.UOW.GroupRepository.Update(id);
+            if (User.Identity.IsAuthenticated)
+            {
+                Group group = UnitOfWork.UOW.GroupRepository.Update(id);
             return View(group);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         [HttpPost]
@@ -65,8 +87,15 @@ namespace KinderGarten.Controllers
         [HttpGet]
         public ActionResult Delete(int? id)
         {
-            Group group = UnitOfWork.UOW.GroupRepository.Delete(id);
+            if (User.Identity.IsAuthenticated)
+            {
+                Group group = UnitOfWork.UOW.GroupRepository.Delete(id);
             return View(group);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         [HttpPost]
@@ -77,5 +106,16 @@ namespace KinderGarten.Controllers
             return RedirectToAction(nameof(Read));
         }
 
+
+        [HttpGet]
+        public ActionResult Details(int id)
+        {
+            List<Kid> allKids = UnitOfWork.UOW.KidRepository.GetAll();
+     
+            Group group = UnitOfWork.uow.GroupRepository.GetByID(id);
+            List<Kid> kidsInTheGroup = allKids.Where(x => x.Groups.Id == group.Id).ToList();
+            group.Kids = kidsInTheGroup;
+            return View(group);
+        }
     }
 }
